Decompose print_eviction_and_entry_diagnostics into diffs.rs helpers
Behavior-preserving (net-diff-0) refactor of the ~470-line god-function
print_eviction_and_entry_diagnostics in verify_execution/run.rs. Extracts
cohesive pub(super) helpers into the existing printer home diffs.rs:

- cdp_change_counts(...) — pure; hoists the nested count_changes and folds
  in the upgrade-count loop. Returns the (creates, updates, deletes,
  upgrade_creates, upgrade_updates) tuple.
- print_cdp_upgrade_entries(...) — the "CDP upgrade entries" block, verbatim.
- coalesce_cdp_final_entries(lcm) -> HashMap — pure; hoists the two coalesce
  closures and the V0/V1/V2 match.
- print_cdp_entry_comparison(ctx, header, entries) — the comparison loop kept
  as ONE helper (shared diffs/missing counters and the `diffs >= 20` cap are
  not split; the bucket_list RwLockReadGuard drop ordering and the
  offer_store_lock MutexGuard scope stay inside it). Leaf detail printers
  print_{offer,account,trustline,pool}_diff_detail and format_missing_entry_key.

Every println!/print!/format!/hex::encode/Sha256 moved byte-for-byte
(literals, arg order, slice bounds [..200]/[..16]/[..8], the diffs>=20 cap),
in the same emission order. Mechanical diff review confirms the removed and
added output-bearing line multisets are identical (61/61). Adds 2 pure-helper
unit tests (cdp_change_counts classification + coalesce last-write-wins).

verify-execution is henyey-only (no stellar-core parity impact); zero
observable diagnostic-output change.

Closes #3424
